backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
pydantic
python-dotenv
